Split maintenance_utils into a maintenance package per route
sandbox_file_permissions.py backs fix-file-permissions, template_sync.py backs sync-protected-files.
